requirements.txt
langchain
langchain-core
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
torch
accelerate
sentence-transformers
grandalf